+++ example/src/screens/index.js
@@ -4,8 +4,12 @@ import AgendaScreen from './agenda';
 import CalendarsScreen from './calendars';
 import CalendarsList from './calendarsList';
 import ExpandableCalendar from './expandableCalendar';
+
 import HorizontalCalendarList from './horizontalCalendarList';
 import MenuScreen from './menu';
+
+import TimelineCalendar from './timelineCalendar';
+
 
 export function registerScreens() {
   Navigation.registerComponent('Menu', () => MenuScreen);
@@ -14,5 +18,5 @@ export function registerScreens() {
   Navigation.registerComponent('CalendarsList', () => CalendarsList);
   Navigation.registerComponent('HorizontalCalendarList', () => HorizontalCalendarList);
   Navigation.registerComponent('ExpandableCalendar', () => ExpandableCalendar);
-
+  Navigation.registerComponent('TimelineCalendar', () => TimelineCalendar);
 }